move database folder to server
