+++ drizzle.config.ts
@@ -3,8 +3,8 @@ import { env } from "@/env";
 import type { Config } from "drizzle-kit";
 
 export default {
-  out: "./src/database/migrations",
-  schema: "./src/database/schema/*",
+  out: "./src/server/database/migrations",
+  schema: "./src/server/database/schema/*",
   driver: "mysql2",
   dbCredentials: { connectionString: env.DATABASE_URL },
   breakpoints: true,
