+++ scripts/database/migration.ts
@@ -1,6 +1,6 @@
-import { db } from "@/database";
+import { db } from "@/server/database";
 import { migrate } from "drizzle-orm/planetscale-serverless/migrator";
 
 (async () => {
-  await migrate(db, { migrationsFolder: "./src/database/migrations" });
+  await migrate(db, { migrationsFolder: "./src/server/database/migrations" });
 })().catch(console.error);
